Ensure Spi::full_duplex_transfer uses uint32_t length
Fixes error when compiling node 10.7.0 node-gyp 3.8.0
../src/spi_binding.cc:196:4: error: non-constant-expression cannot be narrowed from type
      'size_t' (aka 'unsigned long') to 'uint32_t' (aka 'unsigned int') in initializer
      list [-Wc++11-narrowing]

+++ src/spi_binding.cc
@@ -193,7 +193,7 @@ void Spi::full_duplex_transfer(
   struct spi_ioc_transfer data = {
 	  (unsigned long)write,
 	  (unsigned long)read,
-	  length,
+	  (uint32_t)length,
 	  speed,
 	  delay, // Still unsure ... just expose to options.
 	  bits
